requirements.txt
pandas>=2.0,<3.0
numpy>=1.24
scikit-learn>=1.3
gigachat>=0.1
fastapi>=0.110.0,<1.0
uvicorn[standard]>=0.27.0,<1.0
python-multipart>=0.0.9,<1.0
jinja2>=3.1.0,<4.0
python-dotenv>=1.0.0,<2.0
